+++ src/ui/app.js
@@ -99,6 +99,8 @@ class App {
       // Update title.
       this.modalTitle.innerText = controller.getTitle();
 
+      this.updateHeaderHeightCustomProperty();
+
       // Update back button
       if (this.historyIndex >= 2 && this.modalBack.classList.contains('rr-back-hidden')) {
         this.modalBack.classList.remove('rr-back-hidden');
@@ -109,6 +111,43 @@ class App {
 
     this.setLoading(true);
     asyncHandler();
+  }
+
+  /**
+   * Updates a custom css property that contains the current height of the header element.
+   * This is necessary, since inside the .rr-modal-container class the height property is calculated
+   * by subtract the header height from 100%. Since headers have different heights based on their
+   * content, this value needs to be determined whenever the view inside the modal changes.
+   */
+  updateHeaderHeightCustomProperty() {
+    // Name of the custom property for the header height
+    const key = '--rr-modal-header-height';
+    // Select header and modal elements
+    const headerEl = document.querySelector('.rr-modal-header');
+    const modalEl = document.getElementById('rr-omni');
+
+    // Only proceed when both elements are available
+    if (!headerEl || !modalEl) {
+      return;
+    }
+
+    // Get the height of the modal header
+    const headerHeight = headerEl.offsetHeight;
+
+    const { style } = modalEl;
+    if (style.getPropertyValue(key) !== headerHeight) {
+      // Apply the value to the custom property at the modal element
+      style.setProperty(key, headerHeight);
+    }
+
+    if (window?.ResizeObserver && !this.headerObserver) {
+      // Create an observer to update the custom property when the header size changes
+      this.headerObserver = new ResizeObserver(() => {
+        this.updateHeaderHeightCustomProperty();
+      });
+
+      this.headerObserver.observe(headerEl);
+    }
   }
 
   /**
@@ -230,6 +269,10 @@ class App {
 
     // Re-enable document scrolling
     document.querySelector('body').classList.remove('rr-modal-backdrop-body-fix');
+
+    if (this.headerObserver) {
+      this.headerObserver.disconnect();
+    }
   }
 
   addController(controller) {